+++ CHANGELOG.md
@@ -39,7 +39,8 @@ Here is a template for new release sections
 - Sorted parameters in csv´s withing the input folder (#374)
 - Change relative folder path to absolute in tests files (#396)
 - Replace all variables wacc, discount_factor and project_lifetime in the project (#383)
-- LIST_OF_NEW_PARAMETERS can now be edited within `constants.py` (moved from A1) (#384). 
+- `LIST_OF_NEW_PARAMETERS` renamed `EXTRA_CSV_PARAMETERS` and moved from `A1` to `constants.py
+` (#384)
 - Order of parameters in tests/inputs, fixed missing parameters  (#384)
 
 ### Removed